requirements.txt
appwrite
shortuuid
flask
python-dotenv
argon2-cffi